pyproject.toml
[tool.poetry]
name = "transformers-for-vision"
version = "0.1.0"
description = ""
authors = ["Renjith <renjithbaby23@gmail.com>"]
readme = "README.md"
packages = [{include = "transformers_for_vision"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.1"
torchvision = "^0.14.1"
numpy = "^1.24.2"
pre-commit = "^3.0.4"
scikit-learn = "^1.2.1"
pytest = "^7.2.1"
matplotlib = "^3.6.3"
einops = "^0.6.0"
torchmetrics = "^0.11.1"
hydra-core = "^1.3.1"
optuna = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
